requirements.txt
streamlit>=1.35.0
google-genai>=1.0.0
PyMuPDF>=1.24.0
python-docx>=1.1.0
plotly>=5.22.0
python-dotenv>=1.0.0
